refactor: BoxForm 컴포넌트 분리
