+++ src/components/molecules/index.ts
@@ -1,3 +1,4 @@
+export { default as BoxForm } from './BoxForm';
 export { default as Controller } from './Controller';
 export { default as FormInput } from './FormInput';
 export { default as Header } from './Header';
